miam/requirements.txt
Flask==2.1.2
pymongo==3.12.0
tensorflow==2.5.0
twilio==6.57.0
python-dotenv==0.19.2
